kepler-mercado-projeto/requirements.txt
flask
python-oracledb